requirements.txt
Django==4.0.4
celery==5.2.7
django-filter==21.1
django-allauth==0.51.0
python-dotenv==0.20.0
pypugjs==5.9.11